`fn sample_lut`: Make `{sub,b}{x,y}` vars `usize`s.

+++ src/filmgrain.rs
@@ -398,13 +398,13 @@ fn sample_lut<BD: BitDepth>(
     x: c_int,
     y: c_int,
 ) -> BD::Entry {
-    let [subx, suby, bx, by] = [is_subx, is_suby, is_bx, is_by].map(|it| it as c_int);
+    let [subx, suby, bx, by] = [is_subx, is_suby, is_bx, is_by].map(|it| it as usize);
 
-    let randval = offsets[bx as usize][by as usize];
+    let randval = offsets[bx][by];
     let offx = 3 + (2 >> subx) * (3 + (randval >> 4));
     let offy = 3 + (2 >> suby) * (3 + (randval & 0xf as c_int));
-    grain_lut[(offy + y + (BLOCK_SIZE as c_int >> suby) * by) as usize]
-        [(offx + x + (BLOCK_SIZE as c_int >> subx) * bx) as usize]
+    grain_lut[(offy + y) as usize + (BLOCK_SIZE >> suby) * by]
+        [(offx + x) as usize + (BLOCK_SIZE >> subx) * bx]
 }
 
 unsafe extern "C" fn fgy_32x32xn_c_erased<BD: BitDepth>(
